// src/gen/bikehub/model/bicyclesPage.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Bicycle } from './bicycle';


export interface BicyclesPage {
    totalItems?: number;
    currentPage?: number;
    totalPages?: number;
    results?: Array<Bicycle>;
}