// src/gen/bikehub/model/bicycleStatus.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum BicycleStatus {

    Available = 'AVAILABLE',

    Rented = 'RENTED',

    UnderMaintenance = 'UNDER_MAINTENANCE'
}