// src/gen/bikehub/model/createBicycle.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BicycleCategory } from './bicycleCategory';


export interface CreateBicycle {
    currentParkingLotName: string;
    categories: Array<BicycleCategory>;
    chassisId: string;
    brand: string;
    model: string;
    hourlyPrice: number;
}